+++ Source/Debug/AppConfigurationViewController.swift
@@ -102,7 +102,8 @@ class AppConfigurationViewController: DebugTableViewController {
                 [unowned self] cell in
                 self.confirm(
                     from: cell,
-                    message: "Are you sure you want to delete this secret and configuration?  Unless you have copied the secret somewhere you will not be able to restore this identity.",
+                    message: "Are you sure you want to delete this secret and configuration?  Unless you have copied " +
+                    "the secret somewhere you will not be able to restore this identity.",
                     confirmTitle: "Delete"
                 ) {
                     AppConfigurations.delete(self.configuration)
